% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfdQC-package.R
\docType{package}
\name{sfdQC-package}
\alias{sfdQC-package}
\alias{sfdQC}
\title{Functions to support the quallity control of fishery VMS (vessel monitoring system) data.}
\description{
Functions to support the quallity control of fishery
VMS (vessel monitoring system) data.
}
\details{

}
\references{
ICES Working Group on Spatial Fisheries Data: \url{http://www.ices.dk/community/groups/Pages/WGSFD.aspx}.
}
\author{
Colin Millar, Einar Hjorleifsson, Niels Hintzen and Josefine Egekvist.
}
